requirements.txt
pandas==2.0.3
numpy==1.24.3
streamlit==1.28.1
plotly==5.17.0
python-dotenv==1.0.0
# google-generativeai==0.3.0
# openai==1.3.0
# anthropic==0.7.0
groq==0.9.0
faker==19.6.2
pytest==7.4.2